Create BoadrdController and replace rendering functional in controller
